1) Requested by ensemble group, have the WRFDA code work with real size = 4, no impact on the results with eal size = 8 compilation.
M      var/gen_be/gen_be_stage2.f90
M      var/gen_be/gen_be_stage3.f90
M      var/gen_be/gen_be_vertloc.f90
M      var/gen_be/gen_be_stage2_1dvar.f90
M      var/da/da_recursive_filter/da_calculate_rf_factors.inc
M      var/da/da_recursive_filter/da_perform_2drf.inc
M      var/da/da_recursive_filter/da_recursive_filter_1d_adj.inc
M      var/da/da_recursive_filter/da_recursive_filter_1d.inc
M      var/da/da_recursive_filter/da_transform_through_rf_adj.inc
M      var/da/da_recursive_filter/da_transform_through_rf.inc
M      var/da/da_spectral/da_test_spectral.inc
M      var/da/da_spectral/da_vtovv_spectral.inc
M      var/da/da_spectral/da_vtovv_spectral_adj.inc
M      var/da/da_spectral/da_calc_power_spectrum.inc
M      var/da/da_spectral/da_apply_power.inc
M      var/da/da_vtox_transforms/da_transform_vvtovp_adj.inc
M      var/da/da_vtox_transforms/da_transform_vptovv.inc
M      var/da/da_vtox_transforms/da_transform_vvtovp.inc
M      var/da/da_vtox_transforms/da_check_eof_decomposition.inc
M      var/da/da_gen_be/da_filter_regcoeffs.inc
M      var/da/da_control/da_control.f90
M      var/da/da_grid_definitions/da_earth_2_model_wind.inc
M      var/da/da_define_structures/da_allocate_background_errors.inc
M      var/da/da_define_structures/da_define_structures.f90
M      var/da/da_tools/da_1d_eigendecomposition.inc
M      var/da/da_tools/da_get_time_slots.inc
M      var/da/da_tools/da_eof_decomposition_test.inc
M      var/da/da_tools/da_eof_decomposition.inc
M      var/da/da_tools/da_array_print.inc
M      var/da/da_tools/da_get_julian_time.inc
M      var/da/da_minimisation/da_minimise_lz.inc
M      var/da/da_minimisation/da_sensitivity.inc
M      var/da/da_setup_structures/da_setup_be_regional.inc
M      var/da/da_setup_structures/da_setup_be_nmm_regional.inc
M      var/da/da_setup_structures/da_get_bins_info.inc
M      var/da/da_setup_structures/da_truncate_spectra.inc
M      var/da/da_setup_structures/da_rescale_background_errors.inc
M      var/da/da_setup_structures/da_get_vertical_truncation.inc
M      var/da/da_setup_structures/da_scale_background_errors.inc
M      var/da/da_setup_structures/da_setup_be_global.inc
M      var/da/da_setup_structures/da_interpolate_regcoeff.inc
M      var/da/da_obs_io/da_read_obs_bufrgpsro.inc
M      var/da/da_obs_io/da_read_obs_ascii.inc
M      var/da/da_obs_io/da_read_obs_bufr.inc
M      var/da/da_obs_io/da_scan_obs_ascii.inc
M      var/da/da_par_util/da_proc_maxmin_combine.inc
M      var/da/da_par_util/da_proc_stats_combine.inc
M      var/da/da_biascorr_airmass/da_bias_scan.f90
M      var/da/da_biascorr_airmass/da_bias_verif.f90
M      var/da/da_biascorr_airmass/da_bias_airmass.f90
M      var/da/da_main/da_solve.inc
M      var/da/da_etkf/da_innerprod.inc
M      var/da/da_etkf/da_solve_etkf.inc
M      var/da/da_varbc/da_varbc_precond.inc
M      var/da/da_radiance/da_detsurtyp.inc
M      var/da/da_radiance/da_get_innov_vector_rttov.inc
M      var/da/da_radiance/da_rttov_direct.inc
M      var/da/da_radiance/da_read_obs_bufrairs.inc
M      var/da/da_radiance/da_rttov_tl.inc
M      var/da/da_radiance/da_rttov_ad.inc
M      var/da/da_radiance/da_transform_xtoy_crtm_adj.inc
M      var/da/da_radiance/da_read_obs_bufrtovs.inc
M      var/da/da_radiance/da_read_obs_bufrssmis.inc
M      var/da/da_radiance/da_transform_xtoy_crtm.inc
M      var/da/da_radiance/da_cld_eff_radius.inc
M      var/da/da_radiance/da_get_innov_vector_crtm.inc
M      var/da/da_radiance/da_rttov_init.inc
M      compile

2) cleanup Registry.wrfvar, consistent with Resistry.EM

M      Registry/Registry.wrfvar

3) fix from Yongrun for ascii GPSRO data, which rejects GPSRO data above 30km.


M      var/da/da_obs_io/da_read_obs_ascii.inc


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@4173 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d

+++ var/da/da_control/da_control.f90
@@ -611,7 +611,7 @@ module da_control
    real, parameter :: ppbot(npres_print) = (/ 1200.0, 999.9, 899.9, 799.0, 599.9, 399.9,  &
                       299.9,  249.9, 199.9, 149.9, 99.9, 2000./)
 
-   real, allocatable :: time_slots(:)
+   real*8, allocatable :: time_slots(:)
 
    logical :: global
